Move functions, procedures and operators to expressions
