+++ packages/graphql/src/translate/cypher-builder/clauses/sub-clauses/Set.ts
@@ -19,7 +19,7 @@
 
 import type { CypherASTNode } from "../../CypherASTNode";
 import type { CypherEnvironment } from "../../Environment";
-import type { PropertyRef } from "../../PropertyRef";
+import type { PropertyRef } from "../../expressions/PropertyRef";
 import type { Expr } from "../../types";
 import { padBlock } from "../../utils/utils";
 import { SubClause } from "./SubClause";
