+++ packages/graphql/src/translate/cypher-builder/expressions/functions/CypherFunctions.test.ts
@@ -17,8 +17,8 @@
  * limitations under the License.
  */
 
-import { TestClause } from "../utils/TestClause";
-import * as CypherBuilder from "../CypherBuilder";
+import { TestClause } from "../../utils/TestClause";
+import * as CypherBuilder from "../../CypherBuilder";
 
 describe("Functions", () => {
     test("coalesce", () => {
