+++ packages/graphql/src/translate/cypher-builder/expressions/procedures/apoc/RunFirstColumn.ts
@@ -17,11 +17,10 @@
  * limitations under the License.
  */
 
-import { CypherASTNode } from "../../CypherASTNode";
-import { Literal } from "../../variables/Literal";
-import type { Clause } from "../../clauses/Clause";
-import type { Variable } from "../../variables/Variable";
-import type { CypherEnvironment } from "../../Environment";
+import { CypherASTNode } from "../../../CypherASTNode";
+import type { Clause } from "../../../clauses/Clause";
+import type { Variable } from "../../../variables/Variable";
+import type { CypherEnvironment } from "../../../Environment";
 
 export class RunFirstColumn extends CypherASTNode {
     private innerClause: Clause;
