+++ packages/graphql/src/translate/cypher-builder/types.ts
@@ -19,13 +19,13 @@
 
 import type { Variable } from "./variables/Variable";
 import type { Operation } from "./operations/Operation";
-import type { PropertyRef } from "./PropertyRef";
+import type { PropertyRef } from "./expressions/PropertyRef";
 import type { CypherFunction } from "./functions/CypherFunction";
 import type { Literal } from "./variables/Literal";
 import type { Exists } from "./expressions/Exists";
 import type { CypherEnvironment } from "./Environment";
 import type { ComprehensionExpr } from "./expressions/list/ComprehensionExpr";
-import type { MapExpr } from "./variables/map/MapExpr";
+import type { MapExpr } from "./expressions/map/MapExpr";
 import type { BooleanOp } from "./operations/boolean";
 import type { ComparisonOp } from "./operations/comparison";
 import type { RawCypher } from "./clauses/RawCypher";
